more fix

+++ cpp/src/barretenberg/stdlib/primitives/field/field.test.cpp
@@ -991,12 +991,12 @@ template <typename Composer> class stdlib_field : public testing::Test {
             case 1: {
                 a_val = static_cast<uint8_t>(
                     a_val - static_cast<uint8_t>((a_val == 255) ? static_cast<uint8_t>(1) : static_cast<uint8_t>(0)));
-                b_val = a_val + 1;
+                b_val = a_val + static_cast<uint8_t>(1);
                 break;
             }
             case 2: {
                 a_val += static_cast<uint8_t>((a_val == 0) ? static_cast<uint8_t>(1) : static_cast<uint8_t>(0));
-                b_val = a_val - 1;
+                b_val = a_val - static_cast<uint8_t>(1);
                 break;
             }
             default: {
